[Feat] TermsView create
